Delete index.html
Even though root is routed to '/movies' this page is still rendering. I read how you had to manually delete this file in rails 3 so not sure why it's causing me issues since this is rails 4. But after deleting this file then the root URI is routed to the specified view ('/movies').